+++ components/src/Button/IconicButton.js
@@ -29,7 +29,7 @@ const WrapperButton = styled.button(space, ({ disabled }) => ({
   alignItems: "center",
   padding: `${theme.space.half} ${theme.space.none}`,
   color: theme.colors.darkBlue,
-  cursor: disabled ? "arrow" : "pointer",
+  cursor: disabled ? "default" : "pointer",
 
   [`${Icon}`]: {
     borderRadius: theme.radii.circle,